requirements.txt
alembic
black
fastapi
pydantic-settings
sqlalchemy[postgresql_asyncpg]
uvicorn[standard]
psycopg2-binary
asyncpg
starlette
python-multipart
sqlalchemy-utils
pyjwt
python-jose
strawberry-graphql[fastapi]
gunicorn
strawberry-sqlalchemy-mapper
httpx
jinja2
python-dotenv
mangum
uuid6
anyio
zipp>=3.19.1
watchfiles==0.21.0
